Get all images in multiple image tweet
1. Change the file name to datetime.

2. Get all images in multiple image tweet.
The first image name will be just the datetime and the followed will add the "_n", which n represent the number.

3. Get thumbnail from video?
In previous version, the api will return some tweet without the media.
After add the "tweet_mode" = "extended" and it work just charm.
Plus, it also get the thumbnail of video.